refactor: remove outdated alerting components and update user route type import
