+++ src/ui/telemetry/widgets/components/parameter/parameter-form/use-parameter-form.ts
@@ -57,21 +57,21 @@ export function useParameterForm({
       let response: any
 
       if (isEditMode && parameter?.id) {
-        response = await TelemetryService.updateParameter(parameter.id, {
-          name: data.name,
-          unitOfMeasure: data.unit,
-          factor: data.factor,
-          offset: data.offset,
-          isActive: true,
-        })
+        // response = await TelemetryService.updateParameter(parameter.id, {
+        //   name: data.name,
+        //   unitOfMeasure: data.unit,
+        //   factor: data.factor,
+        //   offset: data.offset,
+        //   isActive: true,
+        // })
       } else {
-        response = await TelemetryService.createParameter({
-          name: data.name,
-          unitOfMeasure: data.unit,
-          factor: data.factor,
-          offset: data.offset,
-          isActive: true,
-        })
+        // response = await TelemetryService.createParameter({
+        //   name: data.name,
+        //   unitOfMeasure: data.unit,
+        //   factor: data.factor,
+        //   offset: data.offset,
+        //   isActive: true,
+        // })
       }
 
       if (response.isFailure) {
